benchmark: rename bench_orderings to match the module convention
order_bench_orderings was the one workflow module outside the bench_
prefix; it stays live (the order experiment consumes its variant
catalog and apply_order_variant), so it takes the convention. Pure
rename + import lines (and their import-sort fallout).
