+++ benchmarks/architecture_benchmarking_workflow/bench_orchestrator.py
@@ -48,8 +48,8 @@ from .bench_dry_run import (
     generate_mock_raw_outputs,
 )
 from .bench_metricfns import compute_all_metrics
+from .bench_orderings import compose_stepwise_turns_from_route
 from .bench_routes import Route
-from .order_bench_orderings import compose_stepwise_turns_from_route
 
 # =============================================================================
 # HELPERS
